requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
Pillow==10.0.0
matplotlib==3.7.2
streamlit==1.28.0
